// src/util/enum.ts
export enum CRYPTO_NETWORK {
  BTC = "BTC",
  ERC20 = "ERC20",
  TRC20 = "TRC20",
  BEP20 = "BEP20",
  SOLANA = "SOLANA",
  POLYGON = "POLYGON",
  ARBITRUM = "ARBITRUM",
  OPTIMISM = "OPTIMISM",
  AVALANCHE = "AVALANCHE",
  FANTOM = "FANTOM",
  BSC = "BSC",
  CARDANO = "CARDANO",
  POLKADOT = "POLKADOT",
  COSMOS = "COSMOS",
  TERRA = "TERRA",
  NEAR = "NEAR",
  HARMONY = "HARMONY",
  MOONBEAM = "MOONBEAM",
  CRONOS = "CRONOS",
  KCC = "KCC",
  HECO = "HECO",
  XDAI = "XDAI",
  CELO = "CELO",
  ALGORAND = "ALGORAND",
  TEZOS = "TEZOS",
  ELROND = "ELROND",
  KLAYTN = "KLAYTN",
  OKEX = "OKEX",
}

export enum TRANSACTION_STATUS {
  INITIATED = "INITIATED",
  PENDING = "PENDING",
  AWAITING_PAYMENT = "AWAITING_PAYMENT",
  AWAITING_CRYPTO = "AWAITING_CRYPTO",
  COMPLETED = "COMPLETED",
  FAILED = "FAILED",
  EXPIRED = "EXPIRED",
  CANCELLED = "CANCELLED",
  DISPUTED = "DISPUTED",
  PAYMENT_ACCOUNT_CONFIRMED = "PAYMENT_ACCOUNT_CONFIRMED",
  DEPOSIT_DETECTED = "DEPOSIT_DETECTED",
  DEPOSIT_CONFIRMED = "DEPOSIT_CONFIRMED",
  PAYOUT_INITIATED = "PAYOUT_INITIATED",
  PAYOUT_FAILED = "PAYOUT_FAILED",
  PENDING_PAYOUT = "PENDING_PAYOUT",
}